+++ lib/phoenix_live_view.ex
@@ -375,9 +375,11 @@ defmodule Phoenix.LiveView do
       `clientX`, a keydown event's `keyCode`, etc.
 
   The `phx-capture-click` event is just like `phx-click`, but instead of the click event
-  bubbling up to the closest `phx-click` element, event capturing is used, where the
-  events propagate inwards from the clicked element. This is useful when wanting to bind a click
-  events without receiving bubbled events from child UI elements.
+  being displatched to the closest `phx-click` element as it bubbles up through the DOM, the event
+  is dispatched as it propagates from the top of the DOM tree down to the target element. This is
+  useful when wanting to bind click events without receiving bubbled events from child UI elements.
+  Since capturing happens before bubbling, this can also be important for preparing or preventing
+  behaviour that will be applied during the bubbling phase.
 
   ### Focus and Blur Events
 
@@ -756,7 +758,7 @@ defmodule Phoenix.LiveView do
 
   ## Live navigation
 
-  LiveView provides functionality to allow page navitation using the
+  LiveView provides functionality to allow page navigation using the
   [browser's pushState API](https://developer.mozilla.org/en-US/docs/Web/API/History_API).
   With live navigation, the page is updated without a full page reload.
 
@@ -793,6 +795,7 @@ defmodule Phoenix.LiveView do
 
   `live_patch/2`, `live_redirect/2`, `push_redirect/2`, and `push_patch/2`
   only work for LiveViews defined at the router with the `live/3` macro.
+  Once live navigation is triggered, the flash is automatically cleared.
 
   ### `handle_params/3`
 
@@ -862,70 +865,32 @@ defmodule Phoenix.LiveView do
 
   ## Live Layouts
 
-  Your LiveView will be rendered within the layout specified in your Plug pipeline,
-  such as the default app layout. Assigns defined during `mount` of the root LiveView
-  are accessible in the layout, but the app layout is never updated after the initial
-  render. For a live layout, you must specify an additional layout to use with your
-  LiveView. For example, your regular `app.html` template may display a `@new_message_count`
-  notification, like this:
+  Besides the application layout your LiveView is rendered within,
+  LiveView can also have live layouts, that are rendered within the
+  LiveView life-cycle, on every change.
 
-      <!DOCTYPE html>
-      <html lang="en">
-        <head>
-          <title><%= @page_title %></title>
-        </head>
-        <body>
-          <div>
-            <nav>
-              ...
-              Messages (<%= @new_message_count %>)
-            </nav>
-            <%= render @view_module, @view_template, assigns %>
-          </div>
-        </body>
-      </html>
+  For example, you can define a new `live.html.leex` layout with the
+  dynamic content. You must use `@inner_content` where the output
+  of the actual template will be placed at:
 
-  To allow the `@new_message_count` to be be updated by your LiveView, you can
-  move the dynamic content inside a sub-layout, such as `app_web/templates/layout/live.html.leex`.
-
-  First, you would update your `app.html` layout to keep only the barebones HTML
-  structure:
-
-      <!DOCTYPE html>
-      <html lang="en">
-        <head>
-          <title>...</title>
-          <script>...</script>
-        </head>
-        <body>
-          <%= render @view_module, @view_template, assigns %>
-        </body>
-      </html>
-
-  Next, define a new `live.html.leex` layout with the dynamic content,
-  followed by a render of the inner `@live_view_module`:
-
-      <nav>
-        ...
-        Messages (<%= @new_message_count %>)
-      </nav>
+      <p><%= live_flash(@flash, :notice) %></p>
+      <p><%= live_flash(@flash, :error) %></p>
       <%= @inner_content %>
 
-  Finally, update your LiveView to pass the `:layout` option to `use Phoenix.LiveView`:
+  Finally, update your LiveView to pass the `:layout` option to
+  `use Phoenix.LiveView`:
 
       use Phoenix.LiveView, layout: {AppWeb.LayoutView, "live.html"}
 
-  Or alternatively, you can provide the `:layout` dynamically as an option in mount:
+  The live layout given on `use` is only available for the root live
+  view. If you are rendering child live views or if you want to opt-in
+  to a layout only in certain occasions, you can provide the `:layout`
+  as an option in mount:
 
         def mount(_params, _session, socket) do
           socket = assign(socket, new_message_count: 0)
           {:ok, socket, layout: {AppWeb.LayoutView, "live.html"}}
         end
-
-        def handle_info({:new_messages, count}, socket) do
-          {:noreply, assign(socket, new_message_count: count)}
-        end
-      end
 
   *Note*: The layout will be wrapped by the LiveView's `:container` tag.
 
@@ -1431,10 +1396,10 @@ defmodule Phoenix.LiveView do
       %{assigns: %{^key => _}} ->
         socket
 
-      %{private: %{assigned_new: {assigns, keys}}} ->
+      %{private: %{assign_new: {assigns, keys}}} ->
         # It is important to store the keys even if they are not in assigns
         # because maybe the controller doesn't have it but the view does.
-        socket = put_in(socket.private.assigned_new, {assigns, [key | keys]})
+        socket = put_in(socket.private.assign_new, {assigns, [key | keys]})
         Phoenix.LiveView.Utils.force_assign(socket, key, Map.get_lazy(assigns, key, func))
 
       %{} ->