+++ test/support/live_views/layout.ex
@@ -1,3 +1,17 @@
+defmodule Phoenix.LiveViewTest.ParentLayoutLive do
+  use Phoenix.LiveView
+
+  def render(assigns) do
+    ~L"""
+    <%= live_render @socket, Phoenix.LiveViewTest.LayoutLive, session: @session, id: "layout" %>
+    """
+  end
+
+  def mount(_params, session, socket) do
+    {:ok, assign(socket, session: session)}
+  end
+end
+
 defmodule Phoenix.LiveViewTest.LayoutLive do
   use Phoenix.LiveView, layout: {Phoenix.LiveViewTest.LayoutView, "live.html"}
 
@@ -13,8 +27,8 @@ defmodule Phoenix.LiveViewTest.LayoutLive do
     {:noreply, update(socket, :val, &(&1 * 2))}
   end
 
-  defp maybe_put_layout(socket, %{"live_layout" => {mod, template}}) do
-    {:ok, socket, layout: {mod, template}}
+  defp maybe_put_layout(socket, %{"live_layout" => value}) do
+    {:ok, socket, layout: value}
   end
 
   defp maybe_put_layout(socket, _session), do: {:ok, socket}